update
when user opens application it shows welcome message if user is clicked on database then choose database connection config, then form will be shown. if user is cliked on return home in menu it will return to the original welcome page. Also implemented messages, and if file exists value is read by the application if file does not exists, then new object with database connection is created, and inserted into file implementation class which will save json default config to the file. We need to get that data from the file and put it into text fields into the form.