completed database parse into the fields
we are assuming that names for those fields except password can't have special characters like {, since we are removing those simbols with java functions. However it is possible to implement special request if there is some for it, but i think there is no need for it. For our purpose it will be good we need to save those data in file.